changed pieni version to pieni poc version

+++ deploy/remove.sh
@@ -2,8 +2,15 @@
 
 my_dir="$(dirname "$0")"
 
-# switch to kib project
-oc project u210645-pieni
+if [ "_${PIENI_PROJECT}" = "_" ]; then
+    echo "Need to set version with export PIENI_PROJECT"
+    exit 1
+fi
 
-# remove everything from kib
+#-- switch to kib project
+oc project ${PIENI_PROJECT}
+
+#-- remove everything from kib
 oc delete all -l app=pieni-poc-dev
+
+unset PIENI_PROJECT
